test: move the output_public_note_ntx test to its own file
